% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{biconnectedComponents}
\alias{biconnectedComponents}
\title{Biconnected Components}
\usage{
biconnectedComponents(edges, rt, nNode)
}
\arguments{
\item{edges}{The \code{edge} matrix of a \code{phylo} object.}

\item{rt}{The root node of the phylogeney}

\item{nNode}{The number of nodes in the tree}
}
\description{
Find the biconnected components of a phylogeny
}
\details{
Find the biconnected components of a phylogeny

biconnected components
}
