% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkSim.R
\name{incompleteSampling}
\alias{incompleteSampling}
\title{Sample Tips on a Phylogenetic Network}
\usage{
incompleteSampling(net, rho, stochastic = F)
}
\arguments{
\item{net}{An object of class 'evonet.'}

\item{rho}{The sampling probability.}

\item{stochastic}{If stochastic=FALSE then for a network with n tips we sample n*rho tips. If stochastic=TRUE then each tip porbability rho of being sampled.}
}
\value{
net A network with sampled tips
}
\description{
This function samples tips from a network. Both extant and extinct tips are sampled from the network.
}
