% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isTreeBased}
\alias{isTreeBased}
\title{Determine whether a network is tree-based}
\usage{
isTreeBased(net)
}
\arguments{
\item{net}{A phylogenetic network of class \code{evonet}.}
}
\value{
A logical that is \code{TRUE} if the network is tree-based
}
\description{
This function determines whether a network is tree-based
}
\details{
A phylogenetic network is said to be tree-based if it can be constructed with a base tree that has additional linking arcs added.
}
\references{
See @jetten2016 Corollary 2.11 for the algorithm used to determine whether the network is tree-based
}
