% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.functions.R
\name{make.categorical.draw}
\alias{make.categorical.draw}
\title{Function that makes a draw from a categorical distribution}
\usage{
make.categorical.draw(inheritances, weights)
}
\arguments{
\item{inheritances}{A vector of inheritance probabilities}

\item{weights}{A vector of weights for each inheritance probability}
}
\value{
A function that makes draws from a categorical distribution.
}
\description{
Create a function that makes draws from a categorical distribution. Function calls \code{sample(x=inheritances,size=1,prob=weights)}
}
