% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.functions.R
\name{make.polyploid.model}
\alias{make.polyploid.model}
\title{Model for polyploid evolution across the phylogeny}
\usage{
make.polyploid.model(
  initial_states = 1,
  hyb.event.fxn = NULL,
  hyb.compatability.fxn = NULL,
  time.fxn = NULL,
  spec.fxn = NULL
)
}
\arguments{
\item{initial_states}{the initial ploidy state on the phylogeny. if simulating networks with \code{MRCA=TRUE} then a vector of length two will be required.}

\item{hyb.event.fxn}{A function that describes how ploidy changes after hybridization events. See Details for more information}

\item{hyb.compatability.fxn}{A function that describes whether hybridization events can occur between taxa of different ploidy. See Details for more information}

\item{time.fxn}{A function that describes how ploidy changes over time. See Details for more information}

\item{spec.fxn}{A function that describes how ploidy changes after speciation events. See Details for more information}
}
\value{
A model for polyploidy evolution to be used as the \code{polyploid.model} argument in a `sim.bdh function``
}
\description{
Create a model that dictates how ploidy evolves and affects the diversification of the phylogeny. This function creates a list that dictates how ploidy affects hybridizations, how ploidy is inherited over time, and how ploidy is inherited across speciation and hybridization events.
}
\details{
\code{hyb.event.fxn} is a function that denotes the ploidy of a hybrid child after a hybridization event. The function should have the argument \code{parent_states}, a vector with the trait states of the two parents to the hybrid child. The function should return a single value for the ploidy state of the hybrid child
The default value of \code{NULL} causes the hybrid child to randomly inherit the ploidy of one of the parents.

\code{hyb.compatability.fxn} describes when hybridization events can occur between two taxa based on their ploidy. The function should have the arguments \code{parent_states} and \code{inheritance}. \code{parent_states} is vector with the ploidy states of the hybrid parents while \code{inheritance} is the inheritance probability of the first lineage denoted in \code{parent_states}. The function should return \code{TRUE} for when a hybridization event is allowed to proceed and \code{FALSE} otherwise.
The default value of \code{NULL} causes hybridizations to always occur, regardless of ploidy.

\code{time.fxn} is a function that describes how ploidy changes over time. The function should have the arguments \code{poly_states} and \code{timestep} in that order. \code{poly_states} is a vector containing the ploidy of all taxa while \code{timestep} is the amount of time given for ploidy evolution. The function should return a vector with the updated ploidy states of all taxa.
The default value of \code{NULL} indicates that ploidy will not evolve within a lineage over time. \strong{NOTE:} ploidy can still change at speciation or hybridization events if allowed.

\code{spec.fxn} is a function that describes how ploidy changes at speciation events. The function should have the argument \code{tip_state} which has the state of the lineage just before speciation. The function should return a vector with two values, one denoting the ploidy of each of the two new species after the event.
The default value of \code{NULL} causes the two children lineage to inherit the same ploidy as the parental lineage
}
