% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.functions.R
\name{make.states.incompatible}
\alias{make.states.incompatible}
\title{Incompatability at Hybridization Events}
\usage{
make.states.incompatible()
}
\value{
A function that creates a rule for whether hybridization events occur
}
\description{
Create a function that makes hybridizations unsuccessul if the two parental taxa don't have the same state. The function creates an argument for the \code{hyb.compatability.fxn} element of the \code{trait.model} argument of \code{sim.bdh} style functions.
}
