% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.functions.R
\name{make.uniform.draw}
\alias{make.uniform.draw}
\title{Function that makes a draw from a uniform distribution}
\usage{
make.uniform.draw()
}
\value{
A function that makes draws from a uniform distribution
}
\description{
Create a function that makes draws from a uniform distribution. Function calls \code{runif(1)}
}
\examples{
set.seed(17)
net<-sim.bdh.age(1,1,5,2,1,c(1/3,1/3,1/3),hyb.inher.fxn = make.uniform.draw(),
complete=TRUE)[[1]]
}
