% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.bdh.age.R
\name{sim.bdh.age}
\alias{sim.bdh.age}
\title{Simulate a Phylogenetic Network to a Specififed Number of Taxa}
\usage{
sim.bdh.age(
  age,
  numbsim,
  lambda,
  mu,
  nu,
  hybprops,
  alpha = 1,
  beta = 1,
  frac = 1,
  mrca = FALSE,
  complete = TRUE,
  stochsampling = FALSE
)
}
\arguments{
\item{age}{The time for each simulation.}

\item{numbsim}{Number of networks to simulate.}

\item{lambda}{Speciation rate.}

\item{mu}{Extinction rate.}

\item{nu}{Hybridization rate.}

\item{hybprops}{ector that represents the proportion of Hybridizations that are lineage generative, lineage degenerative, and lineage neutral respectively.}

\item{alpha}{The alpha parameter of the beta distribution. Used to determine inheritance probabilities.}

\item{beta}{The beta parameter of the beta distribution. Used to determine inheritance probabilities.}

\item{frac}{Sampling fraction: The actual number of tips is n/frac, but only n tips are included (incomplete sampling).}

\item{mrca}{If mrca=FALSE: age is the time since origin. If mrca=TRUE: age is the time since most recent common ancestor of the extant tips.}

\item{complete}{If complete = TRUE, the tree with the extinct and non-sampled lineages is returned. If complete = FALSE, the extinct and non-sampled lineages are suppressed.}

\item{stochsampling}{When stochsampling=TRUE: Each tip is included into the final tree with probability frac.}
}
\value{
out Returns a list of numbsim networks with the time since origin / most recent common ancestor being 'age.' If tree goes extinct or no tips are sampled, return value is '0'. If only one extant and no extinct tips are sampled, return value is '1'. Each network has an additional attirbute "inheritance" that represents the inheritance probabilities on the edges in the "reticulation" attribute.
}
\description{
Simulates a Phylogenetic Network under a birth-death-hybridization model. Simulates to a specified ages.
}
