% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.bdh.taxa.R
\name{sim.bdh.taxa}
\alias{sim.bdh.taxa}
\title{Simulate a Phylogenetic Network to a Specififed Number of Taxa}
\usage{
sim.bdh.taxa(
  n,
  numbsim,
  complete = TRUE,
  lambda,
  mu,
  nu,
  hybprops,
  alpha = 1,
  beta = 1,
  frac = 1,
  stochsampling = FALSE
)
}
\arguments{
\item{n}{The number of taxa.}

\item{numbsim}{Number of netowrks to simulate.}

\item{complete}{If complete = TRUE, the tree with the extinct and non-sampled lineages is returned. If complete = FALSE, the extinct and non-sampled lineages are suppressed.}

\item{lambda}{Speciation rate.}

\item{mu}{Extinction Rate}

\item{nu}{Hybridization Rate}

\item{hybprops}{Vector that represents the proportion of Hybridizations that are lineage generative, lineage degenerative, and lineage neutral respectively.}

\item{alpha}{The alpha parameter of the beta distribution. Used to determine inheritance probabilities.}

\item{beta}{The beta parameter of the beta distribution. Used to determine inheritance probabilities.}

\item{frac}{Sampling fraction: The actual number of tips is n/frac, but only n tips are included (incomplete sampling).}

\item{stochsampling}{When stochsampling=TRUE: Each tip is included into the final tree with probability frac.}
}
\value{
out Returns a list of numbsim networks. Each network has an additional attirbute "inheritance" that represents the inheritance probabilities on the edges in the "reticulation" attribute.
}
\description{
Simulates a Phylogenetic Network under a birth-death-hybridization model. Simulates to a specified number of extant tips.
}
