% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_time_sim.R
\name{sim.times.age}
\alias{sim.times.age}
\title{Waiting times of the birth-death-hybridization process}
\usage{
sim.times.age(age, numbsim, lambda, mu, nu, hybprops, startinglineages)
}
\value{
A data frame that has three columns. The first column is the waiting time until the next event. The second column indicates the population size immediately after the event occurs. The last column is \code{1} when the event is a hybridizaiton event and \code{0} otherewise. NOTE: The first row is the root with a waiting time of 0 and the population size being the number of individuals that started the process.
}
\description{
This function simulates a birth-death-hybridization-process and records the waiting times and population size
}
