% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_time_sim.R
\name{waiting.to.phylo.beta}
\alias{waiting.to.phylo.beta}
\title{Waiting times to a phylogeny}
\usage{
waiting.to.phylo.beta(ltt, beta, complete = T)
}
\value{
a phylogenetic network of class \code{evonet}
}
\description{
This function takes the waiting times and population sizes and converts it into a phylogeny
}
