% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isFUstable}
\alias{isFUstable}
\title{Get the level of a Network}
\usage{
isFUstable(net)
}
\arguments{
\item{net}{A phylogenetic network of class \code{evonet}.}
}
\value{
A numeric with the level of the network
}
\description{
This function gets the level of the network
}
