% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isFUstable}
\alias{isFUstable}
\title{Determine whether a phylogeny is FU-stable}
\usage{
isFUstable(net)
}
\arguments{
\item{net}{A phylogenetic network of class \code{evonet}.}
}
\value{
A logical that is \code{TRUE} if the network is FU-stable
}
\description{
This function assesses whether a network is FU-stable
}
\examples{
net<- read.net(text= "((A:7,((B:2,C:2):3)#H1:2::0.6):3,(D:6,#H1:1::0.4):4);")
isFUstable(net) ##returns TRUE
}
