% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.functions.R
\name{make.stepwise}
\alias{make.stepwise}
\title{Make a stepwise decay function}
\usage{
make.stepwise(probs, distances)
}
\arguments{
\item{probs}{Vector of dimension k, where k is the number of different probabilities of success. An individual time between (distances\link{i-1},distances\link{i}] has probability of success prob\link{i}}

\item{distances}{Vector of k, containing the end of each interval where success probabilities shift. The first interval where success is prob\link{1} is [0,distances\link{1}]. For all i>1, the probability of success is prob\link{i} over the interval (distances\link{i-1},distances\link{i}].}
}
\value{
An stepwise decay function
}
\description{
Create a stepwise decay function for genetic distance of two taxa and the probability of success of a hybridization event
}
