% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkSim.R
\name{plottable.net}
\alias{plottable.net}
\title{Create a more Plotting-Friendly phylogenetic Network}
\usage{
plottable.net(net)
}
\arguments{
\item{net}{An object of class \code{evonet}.}
}
\value{
a network to be used with the \code{plot()} function
}
\description{
This function creates a more plotting-friendly \code{evonet} object to be used with the \code{plot()} function
}
